% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(df, yterm, xterms)
}
\arguments{
\item{df}{data frame}

\item{yterm}{y variable name}

\item{xterms}{x variable names}
}
\value{
bootstrapped data frame, summary lm object, betahat estimates, formula used.
}
\description{
This function generates one bootstrap sample
}
\examples{
bootstrap(mtcars,yterm="mpg",xterms="disp+hp")
}
