% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{constructor}
\alias{constructor}
\title{Constructor}
\usage{
constructor(df, yterm, xterms, iter = 1000)
}
\arguments{
\item{df}{data frame for MLR}

\item{yterm}{y variable name}

\item{xterm}{x variable names}
}
\value{
constructor to be used with other methods of this package
}
\description{
This is an object oriented implementation of bootstrap on MLR data. This function creates iter bootstrap samples
}
\details{
Useful to utilize this package when assumptions of MLR are blown. If our sample has size n, the basic process
is take a bunch of samples of size n, with replacement. You can then look at the distribution of sample statistics
like mean and standard deviation. This treats the sample as an estimate of the population. A useful
examination of bootstrapping regression models can be found at
http://statweb.stanford.edu/~owen/courses/305-1314/FoxOnBootingRegInR.pdf
A key quote from that link: The population is to the sample as the sample is to the bootstrap samples.
}
\examples{
obj=constructor(mtcars,yterm="mpg",xterms="disp+hp")
}
