% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{myfun (Title)}
\usage{
myfun(x)
}
\arguments{
\item{x}{this is a double, could be a vector}
}
\value{
squared x
}
\description{
This function is a first attempt (Short Description)
}
\details{
This is really a great function (Full Description)
}
\examples{
myfun(1:10)
}
