% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.DABB5734boot}
\alias{plot.DABB5734boot}
\title{Plot}
\usage{
\method{plot}{DABB5734boot}(x, betaindex = 1, alpha = 0.05)
}
\arguments{
\item{x}{this is the output of the constructor function}

\item{betaindex}{index of beta hat estimate, starts at 0 for beta 0 hat}
}
\value{
plots of beta hat estimates for bootstrap samples
}
\description{
Histogram of beta hat estimates for iter bootstrap samples,
also plots vertical lines for the 100(1-alpha) percent confidence interval
}
\examples{
obj=constructor(mtcars,yterm="mpg",xterms="disp+hp")
plot(obj)
}
