% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.DABB5734boot}
\alias{summary.DABB5734boot}
\title{summary}
\usage{
\method{summary}{DABB5734boot}(x, alpha = 0.05)
}
\arguments{
\item{x}{this is the output of the constructor function}
}
\value{
bootstrap confidence intervals for betahat estimates
}
\description{
Returns confidence intervals for beta hat
}
\details{
confidence intervals are derived using the bootstrap percentile interval given in
http://statweb.stanford.edu/~owen/courses/305-1314/FoxOnBootingRegInR.pdf
Basically we're just taking the quantiles of our iter beta hat estimates
}
\examples{
summary(obj)
}
